Fixed issues and cleaned up radio and gps integration

+++ H750VBT6-flight-board-v0/System/data_transmission.c
@@ -5,7 +5,7 @@
 #include "data_transmission.h"
 
 typedef struct __attribute__((__packed__)) {
-	double	 gps_lat,			gps_long,			gps_alt;
+	float	 gps_lat,			gps_long,			gps_alt;
 	double 	 battery_voltage;
 	uint8_t  pyro_continuity;
 	uint8_t  state;
@@ -27,9 +27,6 @@ void transmitData(SensorData_t* sensorData, FilterData_t* filterData, uint8_t st
 	transmitPacket.state = state;
 
 	//TODO redesign this to support the RX check requirement of the radio, see cc1120_state function in CC1120.c
-
 	// Send packet. A lost packet is not important to send again, so don't worry about success
 	HM_RadioSend((uint8_t*) &transmitPacket, sizeof(transmitPacket));
-
-
 }
